taskmanager/requirements.txt
apiclient==1.0.3
Django==1.11
djangorestframework==3.8.2
google-api-python-client==1.7.3
google-auth-httplib2==0.0.3
httplib2==0.11.3
jsonpickle==0.9.6
oauth2client==4.1.2
requests==2.19.1
slackclient==1.2.1
urllib3==1.23
websocket-client==0.48.0